requirements.txt
argparse
bs4
pillow
requests
